requirements-flink.txt
# Core dependencies
pydantic>=2.4.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Data Processing
apache-flink>=1.17.0
pandas>=2.1.0
numpy>=1.25.0

# Kafka
confluent-kafka>=2.2.0
fastavro>=1.8.0
jsonschema>=4.19.0

# Monitoring
prometheus-client>=0.17.0
opentelemetry-api>=1.19.0

# Security
pyjwt>=2.8.0
cryptography>=41.0.3
